feat(navbar): add conditional buttons for auth state and redirect handling
